Republish roles when role appointments are changed
When creating new roles, appointing a new Role Appointment without first clicking save on the Role resulted in the person appearing on the Organisation page without their Role Title. We are making this change so it will automatically save the Role when new Role Appointments are created or edited.
Additionally uncovered a small issue in Role factory where data wasn't correctly setup, which was previously never called upon.

Co-authored-by: Minno Dang <minno.dang@digital.cabinet-office.gov.uk>