Remove error handling scenario in republishing worker since it is never hit.
We drew a truth table with the different truth values for a combination of `live_edition`, `pre_publication_edition`, `latest_unpublished_edition`, `live_edition.unpublishing`. All cases in the truth table are handled by the if else blocks and the else block where the error is raised is impossible to reach.
The test that previously tested this scenario was triggering a different error down the line (within Withdrawn or Draft), and the message is never matched. The syntax for asserting exception messages is not used like this. The exception needs to be captured from `assert_raised` and then the message of that capture needs to be asserted separately. The string passed to `assert_raise` does nothing.
Removing both test and code block as they are dead code.

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>