Small test refactor for Document Republishing Worker
The test in question was previously stubbing Publishing API calls (not all of them) and forcing them to raise, and asserting no exceptions raised as way to test the methods are not called. We are refactoring this to make it clearer and more conventional that the methods are not expected to be called. We also listed an exhaustive list of external class calls and ensuring they are not called.

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>