Flatten edition links for Editionable Worldwide Organisations
As edition links do not support multi-level or recursive link expansion,
we need to explicitly include all the sub-links that we require to
render this page.

Whilst most of this code is the same as used in the Worldwide
Organisation presenter, it has not been DRY-ed up as that will make it
more difficult to delete the non-editionable version of the presenter
when we've fully migrated Worldwide Organisations to be editionable.