Add association between document and CIPs
For now, the simplest way of associating an edition (corporate information
page) with another edition (editionable worldwide organisation) is to
associate it with its document. This means that users can add and edit
corporate information pages no matter what state the owning worldwide
organisation is in.

This adds an association between corporate information pages and document,
using a join table to avoid adding more columns to the document or edition
tables