Make the CIP pages work for editionable worldwide organisations
This modifies the existing corporate information controller and model so that
they are compatible with editionable worldwide organisations as well as legacy
worldwide organisations.

The owning organisation of a CIP must be an editionable worldwide organisation
and is used for the base path of the CIP, we fetch the live edition if
available, or the most recent version otherwise.

Note: When the editionable worldwide organisations feature flag is enabled,
performing actions related to the corporate information pages of legacy
worldwide organisations will not function (and vice versa). This should be okay
as we don't want these features running side-by-side. The feature flag hides
the legacy worldwide organisations for an extra level of safety.