Refactor temporary asset storage to separate class
Moving the temporary asset storage behaviour into a separate class provides a nice boundary to mock in a unit test. The location in the temp directory is painful to build for testing purposes, so we can mock the directory's storage method instead. This also reduces the cognitive load when reading the asset manager storage class.

The reason the temporary asset storage is required is that when the asset manager storage finishes its store routine, the Carrierwave cache file on the local filesystem is deleted. However, if the create asset worker is invoked asynchronously, it will most likely run after the store! method has finished. In that case it will fail because the file will not be present any longer.