Provide a simpler interface to the create asset worker
The CreateAssetWorker perform method accepted a large number of parameters, many of which were optional. Most of the parameters are derived from the uploader object, so it doesn't make sense to do that derivation in the storage class. Instead we define a method on the worker class to derive the correct parameters. Next we will expose each of the required parameters as methods on the uploader class itself, but we can't do that yet because the storage class tests mock the uploaders excessively so extending its interface and changing the code at the same time would risk introducing a regression.