Refactor create asset worker test
It would be desirable to refactor the create asset worker so that we would need to pass just the assetable class and ID into the worker, and then load the necessary related models from the assetable. Currently, we pass the attachable ID and class separately even though the attachable can be easily fetched from the assetable. Prior to this commit it was hard to refactor the worker to pass just the assetable because the test setup did not relate the attachable to the assetable in any way.

The test also had a lot of shared state created during the setup callback that was either not used in many of the tests or was overridden. This refactor inlines most of the setup code into the tests for greater clarity.