Added previewable and default storage classes
The asset manager storage was a complicated and confusing thing, in large part because it is trying to deal with any and all sorts of asset that might be thrown at it. Most of the time there is no attachable to worry about, and in many cases auth bypass is also not required. It therefore makes sense to separatethe storage into several simpler classes that are able to make some assumptions about the assets they are working with.