Clone worldwide offices for new editions of a worldwide organisations
When a published edition has a new draft edition created, we must copy the
associations for that edition and associate them with the new edition.

We can either (both options are used in various places):
1. Build new records, leaving the old records intact and associated with the
old editions.
2. Point the old records at the new edition.

I've chosen option 1 as it's in-keeping with editionable worldwide organisation
social media accounts.