Don't render previously published component for worldwide organisations
When editing an editionable worldwide organisation, there is an option to
choose a “First published date”.

This is used for things that may have been published in print before appearing
on GOV.UK (or even pre-date GOV.UK).

This isn’t the case for editionable worldwide organisations, so we should
remove the item from the editing form.