Expose default news image on WorldwideOrganisation
The concept of a default news image is used across different content
types, however the impementation vairies.

We want to make it consistent, here we expose the default news image in
the WorldwideOrganisationPresenter which in turn will allow the
PublishingAPI to treat it as a link, this is how the same attribute is
handled for Organisation.

We first exptract the helper methods to a module and then make the
necessary changes the schema has already been updated to reflect this
change.