Republish ministerial index pages when a minister is updated
We were missing a callback on the `Person` model that would update the
ministers index page when a minister is updated.

We don't need to cover the `create` or `destroy` events, since a newly
created person has no role appointments and it is not possible to delete
a person who has role appointments.