Republish ministerial index pages when a ministerial role is updated
We were missing a callback on the `MinisterialRole` model that would
update the ministers index page when a minister is updated.

We don't need to cover the `create` or `destroy` events, since a newly
role has no role appointments and it is not possible to delete a role
that has role appointments.