Republish related editions when role appointment is changed
When a role appointment which has a role associated with an edition
(e.g. Editionable Worldwide Organisation) is created, updated or
deleted, we need to republish then we need to republish the related
editions.

This is because we include an array of role appointments in the links
for the edition, which may need to be reordered as a result of the role
appointment being changed.