Republish related editions when role is changed
When a role associated with an edition (e.g. Editionable Worldwide
Organisation) is updated, we need to republish then we need to republish
the related editions.

This is because we include an array of role appointments in the links
for the edition, which may need to be reordered as a result of the role
appointment being changed.

We don't need to republish when a role is created, as it won't have any
role appointments at that point. Similarly, we don't need to republish
when a role is deleted, as roles can only be deleted if they have no
role appointments.