Extract default new image presentation methods
Both Organisation and WorldwideOrganisation use default_news_image, but
work in different ways. We want to make them consistent.

Here we extract the two methods that are used in the preseter into a
module so they can be resused.

We also rename the methods to help distihguish between the methods on
the internal representation of the object, Organisation here, and the
presentation method.

We have also added some additional coverage around the behaviour.