Denormalize access limitation on to attachment data records
Trying to maintain a reference to a specific attachable from an attachment data model across several instances of the application has resulted in us having to pass that reference through numerous function calls, including in our create asset worker signature.

The only reason we needed to maintain that reference was to derive the organisation IDs that we should pass to Asset Manager so that Asset Manager could restrict access to the attachment asset to users from the organisation that owns the edition.

Denormalizing the data structure so that the organisation IDs are stored as part of the attachment data model will enable us to remove a lot of complexity from the asset handling code.

I think this is safe to do as long as we make sure we update the organisations in Asset Manager when the edition is updated. By recording the organisations at the time of the user action, we maintain the integrity of the transaction.