Present corporate information pages as parts
As part of the larger piece of work to make worldwide organisations editionable,
we want to implement a type of "multi-part" content for worldwide organisation
corporate information pages.

This means that corporate information pages associated with worldwide
organisations will no longer have their own content items, they will instead be
part of the worldwide organisation itself.

This benefits us as we will not have to keep the state of worldwide
organisations and corporate information pages in sync, allowing us to simplify
worldwide organisations in Whitehall.

As part of our migration plan from worldwide organisations to editionable
worldwide organisations, we intend to compare content items to ensure parity.
Therefore, this updates the legacy worldwide organisation to add the required
pages array.