Empty office links for worldwide organisations
Offices are now being included in the main details of the content item,
to allow us to use the same content item to render all pages.

We therefore don't need these pages to be included in links, so need to
empty them.

In a later PR (once all worldwide organisations are republished), we
will remove these links entirely.