Only surface image when not using default news image
Unless a user has added a specific news image for a news article, then a
default image is retrieved from an associated organisation or worldwide
organisation.

Currently, the default image is surfaced as an `image` in the details of a news
story. However, this is problematic because the default image does not update
through dependency resolution when the image is changed on the associated
organisation or worldwide organisation.

This sets the image as nil if the article has a default image. Government
Frontend will render the default news image when the details image is absent
(alphagov/government-frontend#3096), allowing us to replicate the functionality
we are removing here through link expansion instead.