Only show email notifications tab where override is set
The email notifications tab was added to document collections, so that
the notifications could be changed between page level and topic level.

The department have now changed the pages that need changes to the notifications
so this tab can be hidden.

However since the pages that have topic level notificaiton have a different behaviour
we are reatiaing the email notificaitons tab on those pages to inform users and
developers why there is a difference in the emails.