Add migration to unpublish unpublished documents from Publishing API
A small number of documents were previously unpublished but the 'gone'
request was never sent to Publishing API.

There appear to be two reasons for this:
- translations of World Location News pages weren't unpublished when the
  translation was deleted (this was fixed in
  https://github.com/alphagov/whitehall/commit/7faf776275d8d51e4b70c304fad26bd55cc9387d,
  but there were still three documents where the unpublishing request
  hadn't been sent to Publishing API).
- five documents were somehow reverted back to draft state, but with the
  previously published edition also having been deleted; this means
  there is no 'unpublishing' record for the previously published
  edition.

In both cases, it means a bulk republishing results in the wrong data
(with a placeholder schema and render app of `whitehall_admin`) that is
held in the Publishing API database being republished to Content Store.

Therefore sending a request to Publishing API that these should be
deleted properly.