Publish offices and their contacts in the same way as HTML attachments
Worldwide Offices are part of the editionable workflow, so need to be
published/unpublished on certain triggers.

The HTML attachment workflow already offers this functionality, so we
are extending this to work for Worldwide Offices and their Contact.

In later work, this will be extended to include the Worldwide Corporate
Information Pages.