Fix editionable worldwide organisations not being searchable
When testing editionable worldwide organisations in the live environments, we
saw the error `EditionableWorldwideOrganisation is not a searchable class`.

It's likely this is because the `edition_classes` method remains cached despite
switching the Flipflop editionable_worldwide_organisation feature flag.

`EditionableWorldwideOrganisation` is conditional in `edition_classes` only to
prevent it appearing in the search filter when the feature flag is disabled.

Move the condition to the helper instead and add some missing test coverage.