Enable adding mainstream guide sub-pages to document collections
This fixes an error that users would see when trying to add a mainstream guide sub-page to a document collection. The GovukUrl class, unlike the GovukUrlValidator class, did not allow for sub-page paths when verifying that a content item existed at the path provided. We refactor the validator to check the record can access a content item rather than fetching the content item inside the validator so that we can avoid duplicating the behaviour.