Do not present non-edition links for Worldwide Offices
In https://github.com/alphagov/whitehall/pull/8908, we switched
Worldwide Offices to use edition links, so the draft version of the
document can have different links to the parent.

Now that all Worldwide Offices have been republished with edition links,
we can remove the code to publish the non-edition links.