Test that case studies publish editionable worldwide organisations
We are currently migrating worldwide organisations to "editionable worldwide
organisations".

We must include the new editionable worldwide organisations where the old
worldwide organisations were, this includes where other content items have
links to worldwide organisations.

The LinkPresenter was updated in #8912 so the functionality is already there.
This adds test coverage.