Change HTML Attachment public updated at to reflect parent edition last updated
Minor and major version updates previously did not affect HTML attachments. This led to issues where the HTML attachments would appear in latest document search even when the edition was a minor change. With this change, it should reflect the same date that parent edition has for public updated at timestamp.

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>