Remove about us page option for editionable worldwide organisations
Part of the move from using the legacy worldwide organisations to using the new
editionable worldwide organisations involves removing the "about us" corporate
information page.

Instead, this information will live on the editionable worldwide organisation
itself.

This adds methods to the various organisation types to retrieve the types for
that particular organisation.