Load and configure Google Analytics 4
We will run both GA4 and Universal Analytics in tandem while we migrate event tracking to GA4.

Note that we have to accept the default cookie settings on the user's behalf to enable GA4 tracking. This is considered reasonable given that all publishers are part of government.

I have removed the rendering of the Google Tag Manager script component as the GTM script is loaded when initialising GA4