Add Republishing controller with `#index`
This renders the `admin/republishing` template which will allow users to
manually republish a page, with links to verify the page they're going
to be republishing.

We've added a new `:republish` action to `DocumentRules` which requires
the user to have the `gds_admin` role. This is used in the controller to
limit access to the index page, and will be used on other actions
later.

Co-authored-by: Ynda Jas <yndajas@gmail.com>