Add attachments functionality to worldwide organisation page
As worldwide organisation pages are replacing corporate information pages for
editionable worldwide organisations they must be able to have associated
attachments.

Note: Attachments can't be uploaded on the `new` form. This is consistent with
the other models of Whitehall such as corporate information pages and policy
groups.

Note 2: The attachments URL path isn't appended to the full worldwide
organisation pages path. Again, this is consistent with corporate information
pages. We might want to consider making some of the editionable worldwide
organisation paths shallow as a further refinement.