Add "Past prime ministers" republishing flow
This allows GDS admins to republish the "Past prime ministers" page, triggering
the `PresentPageToPublishingApiWorker` job.

We've tried to make this as extendable as possible so we can reuse the `confirm`
action for other republishing tasks in future by adding to the
`republishable_documents` array. We'll probably want to do something similar
for the individual `post` actions to save us having to define an individual
action for each republishable document, but we can do that when we start adding
the new documents in.

Co-authored-by: Ynda Jas <yndajas@gmail.com>