Add audit history for worldwide organisation pages
Worldwide organisation pages are intended to replace corporate information
pages for the new editionable worldwide organisation model.

Whereas corporate information pages are editionable, worldwide organisation
pages are not, and are more similar to worldwide offices.

The upside of this is that they can be published in sync with the editionable
worldwide organisation. One of the downsides is that they won't have their own
audit history.

In order to address this, this adds the pages to the audit history of the
editionable worldwide organisation.

Currently, this simply add conditions to the existing `diff` view. If in future
we want more models to display the audit history of their associated models, we
may want to add the `diff` route to the models themselves and make this more
polymorphic.