Render visual editor for HTML attachments based on feature flag and user permission
We want to release a private-beta version of a new feature, the visual editor.
We only want to share the feature with a select group of users, so we need two layers
of access restrictions, the feature flag for the code and a user permission.

The document types we want to implement the visual editor for are editions (the classic
versioned documents, excluding the latest types that have been editioned) and
HTML attachments. This commit implements the Feature Flag and permission for
HTML attachments.