Remove body and summary from Worldwide Organisation Pages table
In a previous migration, we added fields to the Worldwide Organisation
Pages Translations table that will store the body and summary for the
trnaslations of these pages.

Therefore we don't need these columns in this table anymore.

The data does not need to be migrated over as there are currently no
records of this type in the database.