Add UI for republishing other individual pages
We already added the UI for republishing the "Past Prime Ministers"
page. Since we wrote/refactored this in an DRY-ly extendible way, adding
in other individual presenter-based pages is relatively straightforward

This adds republishing flows for the remaining individual pages that
have dedicated Rake tasks, with a minimal refactor of the
`republishing_pages` method to keep it DRY

In most cases, the `base_path` presenter method was already public, but
a few needed bringing in line, since the controller needs to access this

I've included a feature test for each page since each is presented to
the user as a discrete action. They all use the same slightly abstracted
steps under the hood