Reduce scope of where visual editor is rendered
We are introducing a new feature, the visual editor, behind a feature flag.
Prior work has been done on this, in a more spik-like manner.
For the private-beta release, the scope has been reduced to editions
and HTML attachments, so we are removing previous code where we allowed the
feature to be shown for additional document types.