Add feature tests to cover visual editor rendering
These are skeleton tests to cover the behaviour when feature flag is on,
user permissions are set, and javascript is enabled. Future cards will add
to these tests and complete the journey.

We also added a small test to ensure govspeak editor (textarea) is the default
behaviour when the permission is not set.