Make `WorldwideOrganisationPresenter` properly support drafts
The `WorldwideOrganisationPresenter` has a half implemented solution
where a `state` parameter can be used to present a draft (or other state
version) of the document. This currently doesn't work, as it always
includes the published versions of the corporate information pages.

However, as Worldwide Organisation pages don't support drafts (but their
Coprorate Information Pages do) we never actually use the `state`
paremeter in production, so the broken code never mattered.

Now we want to use the draft version of the Worldwide Organisation's
presented content when doing an integrity check whilst migrating from
non-editionable to editionable Worldwide Organisations.

Therefore fixing the presenter to properly support the state provided as
a parameter.