Copy translations of social media accounts when creating new edition
When creating a new edition of an edition with social media accounts, we
need to also copy the translations.

This is not yet in use, as it is behind the Editionable Worldwide
Organisation feature flag.

We've switched to performing the process after the parent record has
been saved, since we need the record to be persisted before the
translations can be updated.