Return CIPs in a consistent order
At the moment, Corporate Information Pages (CIPs) are grouped by "menu
heading". However, within each group, the CIPs are presented in database
order.

This means they are inconsistently ordered between different
organisations, and between different editions of the same organisation.

Therefore updating this method to return the CIPs in alphabetical order
within the group, to ensure they are always the same.

This will allow us to compare presenter outputs when migrating
worldwide organisations to be editionable.