Add default ordering for World Location associations
MySQL does not guarantee the order in which results are returned,
instead returning results in the order of the record structure on disk.

Normally this is done by ID, but sometimes (e.g. after deletions or
optimisations), records can be in the wrong order.

We want world locations to retain the order that the user added them, so
need to add a default scope that ensures these are always returned in
that order.

A similar default scope is already in use on the `Image` model.