Make worldwide organisation body content consistent
The body of an editionable worldwide organisation without any content
was different to that of a non-editionable worldwide organisation (i.e.
we were including an empty div as govspeak conversion was taking place).

Updating the edtionable presenter to ensure both types are consistent.

Note: the validations have needed to be skipped in the test, since we
don't allow editionable worldwide organisations without a body. However
there are some legacy organisations without a body, who still exist in
the database and could be presented. Therefore skipping validations (and
disabling the rubocop rule) so we can test for those being presented.