Provide a rake task to allow developers to reinstate a deleted document
Sometimes a document gets into a state where its live edition is deleted. We don't currently understand how this happens, but Publishing API and Content Store are not updated to reflect the fact that the edition has been deleted.

It is not possible to recover from this situation via the Whitehall user interface, but the Edition model does expose a method for creating a new draft edition from a deleted edition. This commit provides a rake task to call the method safely so that we can allow users to recover the deleted document.