Fix validation error for edition body when using visual editor
Wrapped the entire editor view in error class when any textarea errors are present.
Pass id to textarea in visual editor view so that the error link in the error summary
can link to & focus on the textarea (if user has permissions but js is disabled).
When the visual editor is rendered (permissions on and js enabled), the id is removed
from the textarea and applied to the contenteditable.

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>
Co-authored-by: Daniel Karaj <daniel.karaj@digital.cabinet-office.gov.uk>