Fix validation errors for HTML attachments when using visual editor
Pass id to textarea in visual editor view so that the error link in the error summary
can link to & focus on the textarea (if user has permissions but js is disabled).
When the visual editor is rendered (permissions on and js enabled), the id is removed
from the textarea and applied to the contenteditable.

Co-authored-by: Daniel Karaj <daniel.karaj@digital.cabinet-office.gov.uk>