Add tests for bulk publishing
Although these are essentially just testing that `true` is passed into
some other API calls, we felt it was important to have this
functionality covered by tests to allow us to refactor the main worker
file with maximum confidence. They're a bit of a pain in that they're
essentially doing the same as their non-bulk counterparts, but this was
the best approach we could come up with!

Co-authored-by: Ynda Jas <yndajas@gmail.com>