Make method name clearer
We found this context slightly unclear when reviewing the `#perform`
method, as it felt like there was a lot of domain knowledge required to
understand why the method would return early here. The "superseded"-ness
of this is hopefully outlined in the test file for this class for any
curious readers.

Given the method name ends with a question mark, we also made the return
type a boolean to make things a little more conventional.

Co-authored-by: Ynda Jas <yndajas@gmail.com>