Extract action list into own method
We'll shortly be adding a new piece of functionality to this worker to
allow for previewing the editions that will be affected when manually
republishing a Document by slug.

By extracting these actions into their own method, we have a single
place to read future actions that will be performed when calling the
worker's `perform` method.

Given that the worker tests were mostly verifying the conditional logic
of the worker, and that conditional logic is now in the Document model,
the tests (or rather the test contexts) have been moved over as well.
Per 56195d5, the integration tests provide assurance that the right API
calls are being made under each condition.

Co-authored-by: George <george@dxw.com>