Add method to show which editions are republished
An upcoming user interface will allow users to enter the slug of a
document to be republished. Republishing a document will kick off a
worker to republish a bunch of editions, and it's useful to see which
editions will be affected before going ahead with republishing them.

This method returns a list of the affected editions, which will later be
presented to the user on the "are you sure you want to republish this?"
page.

Co-authored-by: George <george@dxw.com>