Exit visual editor trial for editions
As part of releasing a private-beta trial for a new visual editor feature,
we are offering the possibility to also exit the trial for a specific document.
The private-beta will only offer limited content editing options, so we
must ensure that users with advanced markdown needs can fall back to using
the standard govspeak markdown editor.

Users in the trial will have a special permission, which has already been implemented.

A user in the trial will always create a new document with the visual editor enabled,
and that document wil remain marked with a DB flag ensuring that all other users with
permissions will continue editing the document in the visual editor.
A user in the trial may exit the visual editor, and that document will remain exited
for all subsequent users. A user outside the trial may also irreversibly exit a document.
Users with JS disabled, be it in the trial or outside the trial, should always have a
seamless experience where they can use the default textarea. Users with JS disabled will
also implicitly exit the document from the trial.

The implementation is done primarily in JS, hiding or showing the visual editor and govspeak
editor respectively. We have used a hidden field whose value we set on the server, and then
modify on the client via JS. The value of the hidden field gets committed on save.

Changes:
- Add a migration to introduce a visual_editor field to editions
- Set the field to true for all editions created by users in the visual editor experiment
- Render it as a hidden field on the edition form defaulting to false (To handle the non-JS case)
- Toggle the field to true when the visual editor is shown (To ensure persistence of the visual editor)
- Introduce a button to exit the visual editor (Displayed only when the visual editor is shown)
- Always render the Govspeak Editor (instead of a backup textarea as before)
- JS to handle toggling of the editor and hidden field when the button is clicked
- The hidden field name is required in the visual editor template to help facilitate extending the behaviour to other
document types (produces a clear error message)
- JS to make sure that the ID is set to either the contenteditable or govspeak textarea as appropriate
to ensure that errors are correctly linked
- Modify the Govspeak Editor JS to get textarea by classname instead of ID to ensure compatibility with previous change

Co-authored-by: Daniel Karaj <daniel.karaj@digital.cabinet-office.gov.uk>