Use V2 of GOV.UK Dependabot Merger
GOV.UK Dependency Merger configuration changed to allow automatic patching of
all dependencies as described in RFC-167 [^1]. Consequently V1 no longer works.
This updates the configuration to use the version 2 convention to reinstate the
basic functionality of automatic patching of internal dependencies only.

Follow up work will be required to enable automatic patching of external
dependencies, which we encourage the team to do.

[^1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-167-auto-patch-dependencies.md