Republish organisation page when contact is updated
A bug was reported in Zendesk recently:
https://govuk.zendesk.com/agent/tickets/5818031

Someone set a contact to appear on the org homepage, but they
never appeared. Running the
`publishing_api:republish:organisation_by_slug` rake task fixes
the issue, so clearly the organisation is not being republished
when it should be.

We already have callbacks for updating the organisation when a
contact is created or destroyed, but not when a contact is
updated. I've been able to consistently reproduce the bug whereby
editing contact info does not get reflected on the org page.
Adding this callback should fix that bug, and should prevent
further issues of this nature.