Exit visual editor trial for HTML attachments
As part of releasing a private-beta trial for a new visual editor feature,
we are offering the possibility to also exit the trial for a specific document,
including HTML attachments.
The private-beta will only offer limited content editing options, so we
must ensure that users with advanced markdown needs can fall back to using
the standard govspeak markdown editor.

Users in the trial will have a special permission, which has already been implemented.

A user in the trial will always create a new document with the visual editor enabled,
and that document wil remain marked with a DB flag ensuring that all other users with
permissions will continue editing the document in the visual editor.
A user in the trial may exit the visual editor, and that document will remain exited
for all subsequent users. A user outside the trial may also irreversibly exit a document.
Users with JS disabled, be it in the trial or outside the trial, should always have a
seamless experience where they can use the default textarea. Users with JS disabled will
also implicitly exit the document from the trial.

The implementation is done primarily in JS, hiding or showing the visual editor and govspeak
editor respectively. We have used a hidden field whose value we set on the server, and then
modify on the client via JS. The value of the hidden field gets committed on save.
This is all mostly already in place from the previous work on editions.

Changes:
- run a migration to add `visual_editor` field to attachments
- allow the new field in the attachments controller
- set the `visual_editor` field as true when building a new attachment
- update the hidden field with the correct name and ID so that it is written to the DB
correctly
- complete feature tests with HTML specific scenarios