Change "refresh" to "republish"
We originally opted for "refresh" as a more generic term that could
encompass republishing, resaving a draft, and updating unpublished and
withdrawn editions. However, the rest of the codebase and documentation
uses the term "republish" to encompass all of this, so while it may not
be the most transparent language, we feel it would be best to remain
consistent for now. In future, it might be useful to revise the language
across the codebase, to improve clarity while maintaining consistency.

Co-authored-by: George Eaton <george@dxw.com>