Save all translations when a draft is saved
This means that when a draft is saved to Publishing API, all
translations of the associated documents are also saved.

Whilst doing this, we noticed that `AttachmentsController` included some
code that was saving the draft to Publishing API itself, then calling
the `draft_updated`, which did the same again. Therefore removing the
duplication.