Create republishing events when republishing pages
This adds a textarea to the Page republishing confirmation page to
capture the reason for republishing a page. This is a mandatory field.

Submitting the form now creates a RepublishingEvent containing:

- the action being performed: e.g. that a page has been scheduled for
  republishing, re-using the original flash message for this
- the reason why the document was republished to help us understand how
  people are using this functionality
- the user who performed the republishing

I've opted not to add a cucumber feature test for the validation error
as although I've seen one or two validation error tests in cucumber
tests, I feel this is well-tested enough in the functional test, and
don't want to double the number of tests as we'd need to add one for
each type of content item anyway.