Fix logic in Worldwide Organisation Presenter
This presenter has untested logic that should produce an empty string if
there is no body text, however we're checking for the return value of
the govspeak helper instead of the body itself.

As the govspeak method always returns some text, it is never possible
for the string to be returned empty.

Therefore updating the logic to reflect how this should work and
including tests to that effect.