Fix bug with social media accounts on new editions
When an Editionable Worldwide Organisation was being editioned (i.e. a
new draft created), the social media accounts were lost, but only for
worldwide organisations that had no translations.

This was because we were using the `build` method to create the record
but never saving it.

For the worldwide organisations with translations, the call to `update`
saved the record, but this never happened for those without
translations.

The tests did not catch this as `create_draft` returns the unsaved
object. Therefore reloading the object in the test to ensure we get the
persisted object from the database.