Remove worldwide organisation field when feature flag enabled
For Editionable Worldwide Organisations, we are assigning roles to them
through the editionable user interface and through the `EditionRole`
relationship.

However the non-editionable version of Worldwide Organisations has them
assigned on the role user interface and through the
`WorldwideOrganisationRole` relationship.

Therefore disabling the non-editionable way of assigning roles when the
editionable worldwide organisation feature flag is enabled.