Shorten display name for Editionable Worldwide Organisations
We currently display the document type as `Editionable worldwide
organisation: Worldwide organisation`, however most of the words are
redundant.

Therefore shortening to `Worldwide organisation`.

This has been done in a generic way, so we can expand the use of single
strings for any other document types that are migrated to be editionable
in the future.